refactor: ResultPage의 Map 컴포넌트에서 상단 헤더를 Header 컴포넌트로 분리
