+++ frontend/src/features/map/components/header/header.styled.ts
@@ -1,0 +1,13 @@
+import { css } from '@emotion/react';
+
+export const base = () => css`
+  position: absolute;
+  width: 100%;
+  top: 20px;
+  padding: 0px 20px;
+  z-index: 10;
+`;
+
+export const top = () => css`
+  width: 100%;
+`;
