+++ frontend/src/pages/components/resultPage/ResultPage.tsx
@@ -5,6 +5,7 @@ import FallBackPage from '@pages/components/fallBackPage/FallBackPage';
 import useSelectedRecommendedLocation from '@pages/hooks/useSelectedLocation';
 
 import BaseLoading from '@features/loading/components/baseLoading/BaseLoading';
+import Header from '@features/map/components/header/Header';
 import Map from '@features/map/components/map/Map';
 import BottomSheet from '@features/recommendation/components/bottomSheet/BottomSheet';
 
@@ -18,7 +19,11 @@ import * as resultPage from './resultPage.styled';
 function ResultPage() {
   const { id } = useParams<{ id: string }>();
   const navigate = useNavigate();
-  const { data: location, isLoading, getRecommendationResult } = useLocationsContext();
+  const {
+    data: location,
+    isLoading,
+    getRecommendationResult,
+  } = useLocationsContext();
 
   const fetchResult = useCallback(async () => {
     try {
@@ -38,7 +43,8 @@ function ResultPage() {
     }
   }, [id, location]);
 
-  const { selectedLocation, changeSelectedLocation } = useSelectedRecommendedLocation();
+  const { selectedLocation, changeSelectedLocation } =
+    useSelectedRecommendedLocation();
 
   const handleSpotClick = (location: RecommendedLocation) => {
     changeSelectedLocation(location);
@@ -50,25 +56,40 @@ function ResultPage() {
 
   if (!location || location.recommendedLocations.length === 0)
     return (
-      <FallBackPage reset={() => navigate('/')} error={new Error('추천 결과가 없습니다.')} text="홈으로 돌아가기" />
+      <FallBackPage
+        reset={() => navigate('/')}
+        error={new Error('추천 결과가 없습니다.')}
+        text="홈으로 돌아가기"
+      />
     );
 
   const { startingPlaces, recommendedLocations } = location;
 
   return (
-    <div css={[flex({ direction: 'column', justify: 'flex-end' }), resultPage.base()]}>
-      <Map
-        startingLocations={startingPlaces}
-        recommendedLocations={recommendedLocations}
+    <div css={resultPage.container()}>
+      <Header
         selectedLocation={selectedLocation}
-        changeSelectedLocation={changeSelectedLocation}
+        onLocationChange={changeSelectedLocation}
       />
-      <BottomSheet
-        startingLocations={location.startingPlaces}
-        recommendedLocations={location.recommendedLocations}
-        selectedLocation={selectedLocation}
-        handleSpotClick={handleSpotClick}
-      />
+      <div
+        css={[
+          flex({ direction: 'column', justify: 'flex-end' }),
+          resultPage.base(),
+        ]}
+      >
+        <Map
+          startingLocations={startingPlaces}
+          recommendedLocations={recommendedLocations}
+          selectedLocation={selectedLocation}
+          changeSelectedLocation={changeSelectedLocation}
+        />
+        <BottomSheet
+          startingLocations={location.startingPlaces}
+          recommendedLocations={location.recommendedLocations}
+          selectedLocation={selectedLocation}
+          handleSpotClick={handleSpotClick}
+        />
+      </div>
     </div>
   );
 }
