+++ frontend/src/pages/components/resultPage/resultPage.styled.ts
@@ -1,6 +1,13 @@
 import { css } from '@emotion/react';
 
+export const container = () => css`
+  position: relative;
+  width: 100%;
+  height: 100%;
+`;
+
 export const base = () => css`
+  position: absolute;
   width: 100%;
   height: 100%;
 `;
